# internal/docs/swagger.yaml
basePath: /api
definitions:
  httpsuite.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  models.AlarmRule:
    properties:
      enabled:
        type: boolean
      id:
        type: string
      last_fired:
        type: string
      last_value:
        type: number
      message:
        type: string
      name:
        type: string
      queue_name:
        type: string
      status:
        $ref: '#/definitions/models.AlarmStatus'
      threshold:
        type: number
      type:
        $ref: '#/definitions/models.AlarmType'
    type: object
...
  models.AlarmStatus:
    enum:
    - ok
    - active
    - inactive
    - firing
    - fired
    - unknown
    type: string
    x-enum-varnames:
    - OK
    - Active
    - Inactive
    - Firing
    - Fired
    - AlarmStatusUnknown
  models.AlarmType:
    enum:
    - channels
    - connections
    - queues
    - unacked
    - queue_messages
    - queue_size
    - no_consumer
    - maintenance
    type: string
    x-enum-varnames:
    - channels
    - connections
    - queues
    - unacked_messages
    - queue_messages
    - queue_size
    - no_consumer
    - maintenance
...
  models.PostRecipient:
    properties:
      email:
        example: ola.normann@normann.no
        type: string
      name:
        example: Slack Channel to team
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.RecipientType'
        enum:
        - webhook
        - email
        example: webhook
      url:
        example: https://hooks.slack.com/services
        type: string
    required:
    - name
    - type
    type: object
...
  models.Recipient:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.RecipientType'
        enum:
        - webhook
        - email
      url:
        type: string
    required:
    - id
    - name
    - type
    type: object
  models.RecipientType:
    enum:
    - webhook
    - email
    - unknown
    type: string
    x-enum-varnames:
    - Webhook
    - Email
    - RecipientTypeUnknown
...
  models.VhostNotification:
    properties:
      name:
        type: string
      notified:
        type: boolean
      recipients:
        items:
          $ref: '#/definitions/models.Recipient'
        type: array
      rules:
        items:
          $ref: '#/definitions/models.AlarmRule'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    name: Norsk helsenett SF
    url: https://github.com/NorskHelsenett/unimq-dashboard
  description: API for RabbitMQ Dashboard application.
  title: RabbitMQ Dashboard API
  version: "1.0"
paths:
...
  /v1/notifications:
    get:
      description: Get all notification vhosts and settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.VhostNotification'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Get Notifications
      tags:
      - Notifications
  /v1/notifications/{vhost-name}:
    delete:
      description: Deletes the notification configuration for a specific vhost
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Delete Notifications
      tags:
      - Notifications
    get:
      description: Get notification rules and settings for a specific vhost
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.VhostNotification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Get Notification on Vhost
      tags:
      - Notifications
  /v1/notifications/{vhost-name}/recipients:
    post:
      consumes:
      - application/json
      description: Add a new notification recipient for a specific vhost
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      - description: Notification Recipient Object
        in: body
        name: recipient
        required: true
        schema:
          $ref: '#/definitions/models.PostRecipient'
      produces:
      - application/json
      responses:
        "201":
          description: Recipient added successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Add a new notification recipient
      tags:
      - Notifications
  /v1/notifications/{vhost-name}/recipients/{recipient-id}:
    delete:
      description: Delete a specific notification recipient for a vhost
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      - description: Recipient ID
        in: path
        name: recipient-id
        required: true
        type: string
      responses:
        "200":
          description: Recipient deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Delete a notification recipient
      tags:
      - Notifications
    get:
      description: Get a specific notification recipient for a vhost
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      - description: Recipient ID
        in: path
        name: recipient-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipient'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Get a notification recipient
      tags:
      - Notifications
...
securityDefinitions:
  OAuth2:
    authorizationUrl: http://localhost:5556/dex/auth
    flow: accessCode
    scopes:
      audience:server:client_id:unimq-dashboard: Issue a token whose audience the
        dashboard API accepts
      email: User email address
      groups: Group membership used for authorization
      openid: OpenID Connect identity
      profile: User profile
    tokenUrl: http://localhost:5556/dex/token
    type: oauth2
  bearer:
    description: '"JWT token for authentication, obtained from the Dex OIDC provider,
      formatted as ''Bearer {token}'' in the Authorization header"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"